demos/demo_S2_AnyAttack/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
open_clip_torch>=2.24.0
timm>=0.9.12
transformers>=4.37.0
accelerate>=0.25.0
webdataset>=0.2.86
img2dataset>=1.41.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
pyarrow>=14.0.0
